+++ tests/test_pipeline/test_pipeline.py
@@ -2,7 +2,6 @@ import re
 from copy import deepcopy
 from datetime import datetime
 from typing import List
-from typing import Tuple
 
 import numpy as np
 import pandas as pd
@@ -22,6 +21,7 @@ from etna.models import SARIMAXModel
 from etna.pipeline import Pipeline
 from etna.transforms import AddConstTransform
 from etna.transforms import DateFlagsTransform
+from tests.utils import DummyMetric
 
 DEFAULT_METRICS = [MAE(mode=MetricAggregationMode.per_segment)]
 
@@ -284,11 +284,11 @@ def test_generate_constant_timeranges_hours():
     (
         (
             False,
-            ["fold_number", "MAE", "MSE", "segment", "SMAPE"],
+            ["fold_number", "MAE", "MSE", "segment", "SMAPE", DummyMetric("per-segment", alpha=0.0).__repr__()],
         ),
         (
             True,
-            ["MAE", "MSE", "segment", "SMAPE"],
+            ["MAE", "MSE", "segment", "SMAPE", DummyMetric("per-segment", alpha=0.0).__repr__()],
         ),
     ),
 )
@@ -299,7 +299,7 @@ def test_get_metrics_interface(
     metrics_df, _, _ = catboost_pipeline.backtest(
         ts=big_daily_example_tsdf,
         aggregate_metrics=aggregate_metrics,
-        metrics=[MAE("per-segment"), MSE("per-segment"), SMAPE("per-segment")],
+        metrics=[MAE("per-segment"), MSE("per-segment"), SMAPE("per-segment"), DummyMetric("per-segment", alpha=0.0)],
     )
     assert sorted(expected_columns) == sorted(metrics_df.columns)
 
@@ -348,44 +348,7 @@ def test_backtest_with_n_jobs(catboost_pipeline: Pipeline, big_example_tsdf: TSD
     assert (forecast_1 == forecast_2).all().all()
 
 
-@pytest.fixture
-def step_ts() -> Tuple[TSDataset, pd.DataFrame, pd.DataFrame]:
-    horizon = 5
-    n_folds = 3
-    train_size = 20
-    start_value = 10.0
-    add_value = 5.0
-    segment = 0
-    timestamp = pd.date_range(start="2020-01-01", periods=train_size + n_folds * horizon, freq="D")
-    target = [start_value] * train_size
-    for i in range(n_folds):
-        target += [target[-1] + add_value] * horizon
-
-    df = pd.DataFrame({"timestamp": timestamp, "target": target, "segment": 0})
-    ts = TSDataset(TSDataset.to_dataset(df), freq="D")
-
-    metrics_df = pd.DataFrame(
-        {"segment": [segment, segment, segment], "MAE": [add_value, add_value, add_value], "fold_number": [0, 1, 2]}
-    )
-
-    timestamp_forecast = timestamp[train_size:]
-    target_forecast = []
-    fold_number_forecast = []
-    for i in range(n_folds):
-        target_forecast += [start_value + i * add_value] * horizon
-        fold_number_forecast += [i] * horizon
-    forecast_df = pd.DataFrame(
-        {"target": target_forecast, "fold_number": fold_number_forecast},
-        index=timestamp_forecast,
-    )
-    forecast_df.columns = pd.MultiIndex.from_product(
-        [[segment], ["target", "fold_number"]], names=("segment", "feature")
-    )
-
-    return ts, metrics_df, forecast_df
-
-
-def test_backtest_forecasts_sanity(step_ts):
+def test_backtest_forecasts_sanity(step_ts: TSDataset):
     """Check that Pipeline.backtest gives correct forecasts according to the simple case."""
     ts, expected_metrics_df, expected_forecast_df = step_ts
     pipeline = Pipeline(model=NaiveModel(), horizon=5)