+++ tests/test_transforms/test_math/test_lag_transform.py
@@ -12,14 +12,6 @@ from etna.transforms.math import LagTransform
 from etna.transforms.math.lags import _OneSegmentLagTransform
 
 
-def equals_with_nans(first_df: pd.DataFrame, second_df: pd.DataFrame) -> bool:
-    """Compare two dataframes with consideration NaN == NaN is true."""
-    if first_df.shape != second_df.shape:
-        return False
-    compare_result = (first_df.isna() & second_df.isna()) | (first_df == second_df)
-    return np.all(compare_result)
-
-
 @pytest.fixture
 def int_df_one_segment() -> pd.DataFrame:
     """Generate dataframe with simple targets for lags check."""
@@ -96,9 +88,8 @@ def test_interface_two_segments_repr(lags: Union[int, Sequence[int]], int_df_two
         assert len(columns_temp) == 1
         generated_column = columns_temp[0]
         assert generated_column == column
-        assert equals_with_nans(
-            df_temp.loc[:, pd.IndexSlice[segments, generated_column]],
-            transformed_df.loc[:, pd.IndexSlice[segments, column]],
+        assert df_temp.loc[:, pd.IndexSlice[segments, generated_column]].equals(
+            transformed_df.loc[:, pd.IndexSlice[segments, column]]
         )